Processing on Notification Service

+++ database/migrations/2026_08_26_112827_create_budget_categories_table.php
@@ -16,7 +16,7 @@ return new class extends Migration
             $table->foreignId('budget_id')->constrained('budgets')->onDelete('cascade');
             $table->foreignId('category_id')->constrained('categories')->onDelete('cascade');
             $table->decimal('limit_amount', 15 , 2);
-            $table->integer('alert_percentage');
+            $table->float('alert_percentage');
             $table->decimal('rollover_amount', 15,2)->nullable()->default(0);
             $table->timestamps();
         });
